Add video controller
