+++ app/src/main/java/me/zsj/interessant/utils/TimeUtils.java
@@ -6,6 +6,9 @@ package me.zsj.interessant.utils;
 
 public class TimeUtils {
 
+    private static final String ZERO = "0";
+    private static final String COLON_SYMBOL = ":";
+
     private TimeUtils() {}
 
     public static int millsToSec(int mills) {
@@ -18,9 +21,10 @@ public class TimeUtils {
         int minute = time / 60 % 60;
         int second = time % 60;
         if (hour == 0) {
-            timeStr = unitFormat(minute) + ":" + unitFormat(second);
+            timeStr = unitFormat(minute) + COLON_SYMBOL + unitFormat(second);
         } else {
-            timeStr = unitFormat(hour) + ":" + unitFormat(minute) + ":" + unitFormat(second);
+            timeStr = unitFormat(hour) + COLON_SYMBOL +
+                    unitFormat(minute) + COLON_SYMBOL + unitFormat(second);
         }
         return timeStr;
     }
@@ -28,9 +32,9 @@ public class TimeUtils {
     private static String unitFormat(int i) {
         String retStr;
         if (i >= 0 && i < 10) {
-            retStr = "0" + Integer.toString(i);
+            retStr = ZERO + Integer.toString(i);
         } else {
-            retStr = "" + i;
+            retStr = String.valueOf(i);
         }
         return retStr;
     }
