+++ app/src/main/res/layout-land/play_activity.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<me.zsj.interessant.widget.VideoController xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/video_controller"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/black">
@@ -57,7 +58,7 @@
                 android:textColor="@android:color/white"
                 android:layout_gravity="center_vertical"
                 android:textSize="12sp"
-                tools:text="00:00"/>
+                android:text="00:00"/>
 
             <SeekBar
                 android:id="@+id/seekBar"
@@ -95,4 +96,4 @@
         app:navigationIcon="?homeAsUpIndicator"
         app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
 
-</FrameLayout>
+</me.zsj.interessant.widget.VideoController>
